+++ geopandas/tests/test_geom_methods.py
@@ -1363,6 +1363,17 @@ class TestGeomMethods:
         with pytest.raises(ValueError):
             original.buffer(distances)
 
+    def test_buffer_distance_series(self):
+        original = GeoSeries([self.p0, self.p0])
+        expected = GeoSeries(
+            [
+                Polygon(((6, 5), (5, 4), (4, 5), (5, 6), (6, 5))),
+                Polygon(((10, 5), (5, 0), (0, 5), (5, 10), (10, 5))),
+            ]
+        )
+        calculated = original.buffer(Series([1, 5]), resolution=1)
+        assert_geoseries_equal(calculated, expected, check_less_precise=True)
+
     def test_buffer_distance_wrong_index(self):
         original = GeoSeries([self.p0, self.p0], index=[0, 1])
         distances = Series(data=[1, 2], index=[99, 98])