+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Bug fixes:
 
+- Fix a regression preventing a Series as an argument for geometric methods (#3363)
 - Support a named datetime index in `explore()` (#3360).
 
 ## Version 1.0.0 (June 24, 2024)